apps/api/requirements.txt
fastapi==0.115.8
uvicorn[standard]==0.34.0
sqlalchemy==2.0.38
psycopg2-binary==2.9.10
alembic==1.14.1
pydantic==2.10.6
pydantic-settings==2.7.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
email-validator==2.2.0
celery==5.4.0
redis==5.2.1
python-multipart==0.0.20
httpx==0.28.1
pytest==7.3.1
pytest-asyncio==0.20.3